refactor: `_create_word()` を移動
